requirements.txt
certifi==2021.5.30
click==8.0.1
ghp-import==2.0.1
gitdb==4.0.7
GitPython==3.1.17
importlib-metadata==4.5.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.1
mkdocs-git-revision-date-plugin==0.3.1
mkdocs-macros-plugin==0.5.5
mkdocs-material==7.1.8
mkdocs-material-extensions==1.0.1
packaging==20.9
Pygments==2.9.0
pymdown-extensions==8.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
pyyaml-env-tag==0.1
six==1.16.0
smmap==4.0.0
termcolor==1.1.0
watchdog==2.1.2
zipp==3.4.1
